% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_VICI.R
\name{boxplot_VICI}
\alias{boxplot_VICI}
\title{Plotting function for displaying boxplots and associated p-values}
\usage{
boxplot_VICI(data_df, pval_2plot, response_name, inter = TRUE,
  baseline = NULL)
}
\arguments{
\item{pval_2plot}{a \code{data.frame} with the p-values to display.}

\item{inter}{a logical flag indicating whether we are in the interarm setting or not.
Default is \code{TRUE}.}

\item{res_2plot}{a \code{data.frame}.}
}
\value{
a \code{ggpubr} plot object
}
\description{
Internal function for displaying significance boxplots
}
\author{
Boris Hejblum
}
\keyword{internal}
