% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_jaclist}
\alias{compute_jaclist}
\title{Compute_jaclist quantities needed for the Satterthwaite
approximation.}
\usage{
compute_jaclist(object, tol = 1e-06)
}
\arguments{
\item{object}{a \code{gls} object.}

\item{tol}{a tolerance}
}
\value{
a list.
}
\description{
Computes vcov of variance parameters (theta, sigma), jacobian of
each variance parameter etc.
}
\details{
This code is adapted from code in compute_auxillary internal 
function of pbkrtest package.
}
\keyword{internal}
