% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BW_functions.R
\name{ddf_BW}
\alias{ddf_BW}
\title{Between-Within functions to obtain Denominator degrees of freedom}
\usage{
ddf_BW(object, L, data, input)
}
\description{
Between-Within functions to obtain Denominator degrees of freedom
}
\keyword{internal}
