% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{devfun_gls}
\alias{devfun_gls}
\title{Compute Full Deviance}
\usage{
devfun_gls(varpar, gls_obj)
}
\arguments{
\item{varpar}{variance parameters.}

\item{gls_obj}{a \code{gls} object.}
}
\value{
the full deviance, a numerical scalar.
}
\description{
Compute Full Deviance
}
\details{
This code is adapted from code in \code{devfun_vp} internal function of 
\pkg{pbkrtest} package.
}
\keyword{internal}
