% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KR_functions.R
\name{get_KR1D}
\alias{get_KR1D}
\title{Functions to calculate Kenward-Roger approximation of ddf}
\usage{
get_KR1D(model, L)
}
\description{
Functions not finished and not used inside package. We don't find the same thing that in SAS. 
The problem may be come of \code{getME.gls} function for the definition of groups to compute
varCov.
}
\keyword{internal}
