% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_VICI.R
\name{heatmap_vici}
\alias{heatmap_vici}
\title{A heatmap function for displaying}
\usage{
heatmap_vici(res_2plot)
}
\arguments{
\item{res_2plot}{a \code{data.frame}}
}
\value{
a \code{ggplot2} plot object
}
\description{
Internal function for displaying significance heatmap when multiple conditions are tested
}
\author{
Boris Hejblum
}
\keyword{internal}
