% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_VICI.R
\name{histogram_VICI}
\alias{histogram_VICI}
\title{Plotting function for displaying histograms and associated p-values}
\usage{
histogram_VICI(
  data_df,
  pval_2plot,
  response_name,
  input,
  inter = TRUE,
  baseline = NULL
)
}
\arguments{
\item{data_df}{a \code{data.frame}.}

\item{pval_2plot}{a \code{data.frame} with the p-values to display.}

\item{response_name}{a character string indicating the name of the response.}

\item{input}{internal input from UI.}

\item{inter}{a logical flag indicating whether we are in the interarm setting or not.
Default is \code{TRUE}.}

\item{baseline}{baseline value used in title when \code{inter} is \code{FALSE}.
Default is \code{NULL}.}
}
\value{
a \code{ggpubr} plot object
}
\description{
Internal function for displaying significance histograms
}
\author{
Clément NERESTAN
}
\keyword{internal}
