% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraarm_fit.R
\name{intraarm_fit}
\alias{intraarm_fit}
\title{Fitting GLS For Intra-Arm Setting}
\usage{
intraarm_fit(transformed_data, tested_time, input)
}
\description{
Fitting GLS For Intra-Arm Setting
}
\keyword{internal}
