% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_modelfit.R
\name{mod_modelfit_ui}
\alias{mod_modelfit_ui}
\alias{mod_modelfit_server}
\title{mod_modelfit_ui and mod_modelfit_server}
\usage{
mod_modelfit_ui(id)

mod_modelfit_server(input, output, session, datas, parent, origin)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{datas}{internal}

\item{parent}{internal}

\item{origin}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
