% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings_pan.R
\name{mod_settings_pan_ui}
\alias{mod_settings_pan_ui}
\alias{mod_settings_pan_server}
\title{mod_settings_pan_ui and mod_settings_pan_server}
\usage{
mod_settings_pan_ui(id)

mod_settings_pan_server(input, output, session, data, parent)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data}{internal}

\item{parent}{}
}
\description{
A shiny Module.
}
\keyword{internal}
