% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDownloadHandlerForPlots.R
\name{myDownloadHandlerForPlots}
\alias{myDownloadHandlerForPlots}
\title{Custom download handler for plots}
\usage{
myDownloadHandlerForPlots(name, plot_obj)
}
\arguments{
\item{name}{output file name}

\item{plot_obj}{a plot object to be downloaded}
}
\value{
a \code{ggpubr} plot object
}
\description{
Custom download handler for plots
}
\author{
Boris Hejblum
}
\keyword{internal}
