% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qform}
\alias{qform}
\title{Utility Functions for VICI Package}
\usage{
qform(L, V)
}
\arguments{
\item{L}{a numeric vector.}

\item{V}{a symmetric numeric matrix.}
}
\value{
a numerical scalar.
}
\description{
Compute Quadratic Form
}
\keyword{internal}
