% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch VICI Shiny App}
\usage{
run_app(host = "0.0.0.0", port = 8080, ...)
}
\arguments{
\item{host}{Default is "127.0.0.1", see \link[shiny]{runApp} for details.}

\item{port}{Default is 3838, see \link[shiny]{runApp} for details.}

\item{...}{additional arguments to be passed to the \link[shiny]{runApp} function.}
}
\description{
Launch VICI Shiny App
}
\examples{
if(interactive()){
vici::run_app()
}

}
