% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{waldCI}
\alias{waldCI}
\title{Compute Wald Confidence Interval}
\usage{
waldCI(estimate, se, df = Inf, level = 0.95)
}
\arguments{
\item{estimate}{an estimated coefficient.}

\item{se}{standard error of \code{estimate}.}

\item{df}{degrees of freedom associate to \code{estimate}. \code{df = Inf} is 
allowed.}

\item{level}{level of confidence interval.}
}
\value{
a matrix of lower and upper confidence interval.
}
\description{
Compute Wald Confidence Interval
}
\details{
This code is greatly inspired by code from the lmerTest
    package.
}
\keyword{internal}
