% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BW_functions.R
\name{ddf_BW}
\alias{ddf_BW}
\title{Between-Within functions to obtain Denominator degrees of freedom}
\usage{
ddf_BW(object, L)
}
\arguments{
\item{object}{a \code{gls} model object.}

\item{L}{a numeric vector of the same length as the number of fixed effects from \code{object}.}
}
\value{
a \code{number} object
}
\description{
Internal function to calculate Between-Within degrees of freedom
}
\keyword{internal}
