% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_ownfunctions.R
\name{get_coefmat_gls}
\alias{get_coefmat_gls}
\title{Functions to obtain coef, ddf, p-value}
\usage{
get_coefmat_gls(
  model,
  ddf = c("Satterthwaite", "Kenward-Roger", "Between-Within")
)
}
\description{
Functions to obtain coef, ddf, p-value
}
\keyword{internal}
