% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_ownfunctions.R
\name{get_coefmat_gls}
\alias{get_coefmat_gls}
\title{Functions to obtain coef, ddf, p-value}
\usage{
get_coefmat_gls(
  model,
  ddf = c("Satterthwaite", "Kenward-Roger", "Between-Within")
)
}
\arguments{
\item{model}{a \code{gls} model.}

\item{ddf}{degrees of freedom approximation.}
}
\value{
a matrix containing coefficient, ddf and p-value
}
\description{
This function allows to calculate the different approximations of degrees of 
freedom and returns the table of results in the app.
}
\keyword{internal}
