% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_VICI.R
\name{heatmap_vici}
\alias{heatmap_vici}
\title{A heatmap function for displaying}
\usage{
heatmap_vici(res_2plot, inter = TRUE, baseline = NULL)
}
\arguments{
\item{res_2plot}{a \code{data.frame}}

\item{inter}{a logical flag indicating whether we are in the interarm setting or not.
Default is \code{TRUE}.}
}
\value{
a \code{ggplot2} plot object
}
\description{
Internal function for displaying significance heatmap when multiple conditions are tested
}
\author{
Boris Hejblum
}
\keyword{internal}
