% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varBetafun_gls}
\alias{varBetafun_gls}
\title{Compute cov(Beta) For a GLS Model}
\usage{
varBetafun_gls(varpar, gls_obj)
}
\arguments{
\item{varpar}{variance parameters.}

\item{gls_obj}{a \code{gls} object.}
}
\value{
cov(Beta), a numerical scalar.
}
\description{
Compute cov(Beta) For a GLS Model
}
\details{
This code is adapted from code in get_covbeta internal function of 
pbkrtest package.
}
\keyword{internal}
